+++ frontend/src/views/Login.tsx
@@ -1,22 +1,26 @@
 import React, { useContext } from "react";
 
-// Import context
-import AuthContext from "../../context/auth-context";
-
 // Import components
-import { auth } from "../../helpers/auth";
-import { Form } from "../organism/Form/Form";
+import { Form } from "../components/organism/Form/Form";
 
+// Import styles
 import "./ConnectionForm.css";
 
+// Import context
+import AuthContext from "../context/auth-context";
+
+// Import auth
+import { auth } from "../helpers/auth";
+
 // TODO: Change path of auth function (in submitHandler)
-export const Register = () => {
+export const Login = () => {
   const user = useContext(AuthContext);
+
   return (
     <section>
       <Form
-        type="register"
-        title="Create an Account"
+        type="login"
+        title="Connect to your Account"
         submitHandler={(data) => auth(data, "users/1", user)}
       />
     </section>
