+++ frontend/src/views/Register.tsx
@@ -1,26 +1,22 @@
 import React, { useContext } from "react";
 
-// Import components
-import { Form } from "../organism/Form/Form";
+// Import context
+import AuthContext from "../context/auth-context";
 
-// Import styles
+// Import components
+import { auth } from "../helpers/auth";
+import { Form } from "../components/organism/Form/Form";
+
 import "./ConnectionForm.css";
 
-// Import context
-import AuthContext from "../../context/auth-context";
-
-// Import auth
-import { auth } from "../../helpers/auth";
-
 // TODO: Change path of auth function (in submitHandler)
-export const Login = () => {
+export const Register = () => {
   const user = useContext(AuthContext);
-
   return (
     <section>
       <Form
-        type="login"
-        title="Connect to your Account"
+        type="register"
+        title="Create an Account"
         submitHandler={(data) => auth(data, "users/1", user)}
       />
     </section>
